requirements.txt
streamlit==1.32.0
bcrypt==4.1.2
Pillow>=10.0.0
pandas>=1.5.0
